% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_estimate_params.R
\name{add_estimate_params}
\alias{add_estimate_params}
\title{Add Estimate Parameters}
\usage{
add_estimate_params(df, parameters)
}
\arguments{
\item{df}{Data frame the estimate (B) parameters will be added to.}

\item{parameters}{Date frame containing the estimate parameters (from "get_RL_model_estimates()")}
}
\value{
A data frame with columns containing the model estimates added to it.
}
\description{
Add estimate parameters to a data frame.
}
\examples{
\dontrun{
model_params <- get_RL_model_estimates()
all_correct \%>\% add_estimate_params(model_params)
}
}
