% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_subj_rand_eff.R
\name{add_subj_rand_eff}
\alias{add_subj_rand_eff}
\title{Add Subject Random Effect}
\usage{
add_subj_rand_eff(df, tau_0)
}
\arguments{
\item{df}{Data frame to add the subject random effect sd to.}

\item{tau_0}{numeric input. Subject random effect SD}
}
\value{
A data frame with a column containing the subject random effect added to it.
}
\description{
Add the subject random effect to a data frame
}
\examples{
\dontrun{
# If the subject random effect SD = 0.5
all_correct \%>\% add_subj_rand_eff_sd(0.5)
}
}
