% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_RL_desc.R
\name{check_RL_desc}
\alias{check_RL_desc}
\title{Check RL descriptive statistics}
\usage{
check_RL_desc(orig, sim)
}
\arguments{
\item{orig}{\code{tibble} original descriptive statistisc. From "get_desc_RL()"}

\item{sim}{\code{tibble} simulated data frame}
}
\value{
\code{tibble}
}
\description{
Creates a table comparing the original and simulated RL descriptive statistics.
}
\examples{
\dontrun{
orig_desc <- get_desc_RL()
check_RL_desc(orig_desc, sim_data) # where sim_data is the simulated data set.
}
}
