% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_spatial_encodings.R
\name{check_spatial_encodings}
\alias{check_spatial_encodings}
\title{Check spatial encodings}
\usage{
check_spatial_encodings(
  group_parameters,
  n_trials = 15,
  sim_df,
  formattable = TRUE
)
}
\arguments{
\item{group_parameters}{\code{tibble} containing the group parameters (from "get_group_params()").}

\item{n_trials}{numeric. number of trials per participant. "n_trials = 15" by default.}

\item{sim_df}{\code{tibble} containing the simulated data frame}

\item{formattable}{Logical. If TRUE format table. "formattable = TRUE" by default.}
}
\value{
\code{tibble}
}
\description{
Check that percentage of correct spatial encodings is expected"
}
\examples{
\dontrun{
check_spatial_encodings(group_parameters = group_params, sim_df = sim_data)
}
}
