% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixed_effects_tab.R
\name{fixed_effects_tab}
\alias{fixed_effects_tab}
\title{Fixed effect table}
\usage{
fixed_effects_tab(mod)
}
\arguments{
\item{mod}{\code{glmerMod}}
}
\value{
\code{formattable}
}
\description{
Table with the input model's fixed effect
}
\examples{
\dontrun{
fixed_effects_tab(sim_mod)
}
}
