% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csv.R
\name{get_csv}
\alias{get_csv}
\title{Get previously simulated data}
\usage{
get_csv(file_name)
}
\arguments{
\item{file_name}{String argument. Name of file.}
}
\value{
Return a \code{tibble}.
}
\description{
Retrieves previously simulated data stored in csv file.
}
\examples{
# extdata is the name of the folder the file in.
path = system.file("extdata", "pre_sim_data.csv", package = "Karadoller2020")
get_simulation(path)
}
