% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_desc_RL.R
\name{get_desc_RL}
\alias{get_desc_RL}
\title{Mean and SD proportions of relational lexemes across age, language and viewpoint}
\usage{
get_desc_RL(kable_format = FALSE)
}
\arguments{
\item{kable_fomart.}{Logical. If TRUE apply kable styling to data frame. "kable_format = FALSE" by default.}
}
\value{
A data frame containing the mean and sd proportions of relational lexemes by participant.
}
\description{
Generates a data frame containing the mean and sd proportions of relational lexemes. Taken from table 3.
}
\examples{
get_desc_RL()
}
