% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_params.R
\name{get_group_params}
\alias{get_group_params}
\title{Get Group Parameters}
\usage{
get_group_params()
}
\value{
Data frame containing the group parameters. Number of participants by age groups and language status and their probability of getting a correct spatial encoding.
}
\description{
Creates a data frame containing the group paramters. These can be found in Table 1 in the paper.
}
\examples{
get_group_params()
}
