% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_head_tail.R
\name{get_head_tail}
\alias{get_head_tail}
\title{Get head and tail}
\usage{
get_head_tail(df, n_head, n_tail)
}
\arguments{
\item{df}{\code{tibble}}

\item{n_head}{numeric argument.}

\item{n_tail}{numeric argument.}
}
\value{
\code{tibble}. First and last N rows of the input data frame.
}
\description{
combined head() and tail()
}
\examples{
# get the first two and last row of my_df
my_df <- tibble(x = c(1,2,3,4),
 y = c("a", "b","c","d"))
my_df \%>\% get_head_tail(n_head = 2, n_tail = 1)
}
