% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_RL_model_estimates.R
\name{get_RL_model_estimates}
\alias{get_RL_model_estimates}
\title{Get RL Model Estimates}
\usage{
get_RL_model_estimates()
}
\value{
A data frame containing the model estimates for the fixed effects from the glmer model for frequency of RL use.
}
\description{
Generates a data frame containing the model estimates (B) for the fixed effects from the glmer model for frequency of RL use. (Table 5)
}
\examples{
get_RL_model_estimates()
}
