% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trial_params.R
\name{get_trial_params}
\alias{get_trial_params}
\title{Get Trial Parameters}
\usage{
get_trial_params()
}
\value{
Data frame containing the trial parameters (number of trials, number of view dependent trials, number of view independent trials)
}
\description{
Creates data frame containing the trial parameters
}
\examples{
get_trial_params()
}
