% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_analysis.R
\name{power_analysis}
\alias{power_analysis}
\title{Power analysis}
\usage{
power_analysis(df, effect_of_interest)
}
\arguments{
\item{df}{\code{tibble} containing model outputs.}

\item{effect_of_interest}{String argument}
}
\value{
numeric value
}
\description{
Conducts a power analysis on a data frame with model outputs
}
\examples{
\dontrun{
power_table \%>\% power_analysis(effect_of_interest = "age_dev")}

}
