
\name{prev_sim_power_tab}
\alias{prev_sim_power_tab}
\title{
Power table
}
\description{
Power to detect a significant effect of age. Contains the fixed effects of 100 model simulations.

}
\usage{
power_table <- Karadoller2020::get_csv("prev_sim_power_tab.csv")
}
\format{A data frame containing 800 rows and 7 variables.


}
\note{RL stands for relational lexeme.}


\docType{data}


