% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RL_violin_plot.R
\name{RL_violin_plot}
\alias{RL_violin_plot}
\title{Create violin plot for RL use}
\usage{
RL_violin_plot(df)
}
\arguments{
\item{df}{\code{tibble}}
}
\value{
A violin plor
}
\description{
Create violin plot for RL use
}
\examples{
sim_data <- sim_df()
RL_violin_plot(sim_data)
}
