% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_df.R
\name{sim_df}
\alias{sim_df}
\title{Simulate the data}
\usage{
sim_df(
  tau_0 = 0.9,
  sensitivity_analysis = FALSE,
  n_participants,
  change_coeff,
  change_coeff_value
)
}
\arguments{
\item{tau_0}{Numeric argument. Subject random effect standard deviation. "tau_0 = 0.9" by default.}

\item{sensitivity_analysis}{Logical. Allows the number of participants and value of one coefficient to be manually overridden if TRUE. "sensitivity_analysis = FALSE" by default.}

\item{n_participants}{Numeric argument. Specify total number of participants. Must be divisible by the number of groups (4 in this case) without a remainder. Ensures that the groups are balanced.}

\item{change_coeff}{String argument. Specifies the changed coefficient for the sensitivity analysis. E.g. "b0","b1". NOT "age" or "age_dev".}

\item{change_coeff_value}{Numeric argument. Value of changed coefficient}
}
\value{
\code{tibble} containing simulated data.
}
\description{
Runs the simulation code for one data frame.
}
\examples{
# Normal version
sim_df() \%>\% head()

# Sensitivity analysis

sim_df(sensitivity_analysis = TRUE,
n_participants = 60, # 60 participants in total (15 in each condition)
change_coef = "b1", # change "b1" (age coef)
change_coef_value = 90 # change "b1" to 90
) \%>\% head()

}
