% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_power_table.R
\name{sim_power_table}
\alias{sim_power_table}
\title{Simulate Power Table}
\usage{
sim_power_table(n_simulations = 100, alpha = 0.05)
}
\arguments{
\item{n_simulations}{Numeric argument. Number of simulations. "n_simulations = 100" by default.}

\item{alpha}{Numeric argument. Alpha value for the power analysis. Acceptable type one error rate. "alpha = 0.05" by default.}
}
\value{
\code{tibble}. Contains the model outputs for N simulations.
}
\description{
Simulate the power table. Simulates and analyses the data N times.
}
\examples{
# Simulate and analyse the data twice.
sim_power_table(2)
}
