% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_RL_from_logodds.R
\name{sim_RL_from_logodds}
\alias{sim_RL_from_logodds}
\title{Simulate RL from log odds}
\usage{
sim_RL_from_logodds(df)
}
\arguments{
\item{df}{data frame}

\item{logodds_col}{name of the log odds column."logodds_col = logodds" by default}
}
\value{
A data frame containing the simulated RL based on the log odds column.
}
\description{
Simulate RL from log odds
}
\examples{
\dontrun{
sim_data \%>\% sim_RL_from_logodds()
}
}
