% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_spatial_encoding.R
\name{sim_spatial_encoding}
\alias{sim_spatial_encoding}
\title{Simulate the Correct Spatial Encoding Data}
\usage{
sim_spatial_encoding(
  group_params,
  trial_params,
  age_ref = "child",
  lang_ref = "late",
  viewpoint_ref = "independent"
)
}
\arguments{
\item{group_params}{data frame input. group_params can be accessed using the "get_group_params()" function}

\item{trial_params}{data frame input. trial_params can be accessed using the "get_trial_params()" function}

\item{age_ref}{string input. "age_ref = child" by default. Picks the reference category.}

\item{lang_ref}{string input. "lang_ref = native" by default. Picks the reference category.}

\item{viewpoint_ref}{string input. "viewpoint_ref = independent" by default. Picks the reference category.}
}
\value{
Returns a data frame containing a simulation of the correct spatial encoding data
}
\description{
Applies "sim_spatial_encoding_single()" to all the rows in group_params and then combines them into one data frame.
}
\examples{
\dontrun{sim_spatial_encoding(group_params, trial_params)}
}
