% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_spatial_encoding_single.R
\name{sim_spatial_encoding_single}
\alias{sim_spatial_encoding_single}
\title{Simulate Spatial Encodings for a Single Group}
\usage{
sim_spatial_encoding_single(
  n_participants,
  p_correct,
  language,
  age,
  n_trials,
  n_view_dep,
  n_view_indep
)
}
\arguments{
\item{n_participants}{Number of participants in the group.}

\item{p_correct}{Probability the participant encodes the correct spatial relation between the figure and ground item.}

\item{language}{Language status. "native" or "late".}

\item{age}{Whether the participants is a child or adult. "adult" or "child.}

\item{n_trials}{Number of trials}

\item{n_view_dep}{Number of view dependent trials.}

\item{n_view_indep}{Number of view independent trials.}
}
\value{
\code{tibble} containing the simulated correct spatial encodings for one group.
}
\description{
Simulates the spatial encoding for a single group (e.g. native children, late adults). spatial_encoding: 1 = correct, 0 = incorrect
}
\examples{
# Simulates spatial encodings for native adults.
sim_spatial_encoding_single(15,.98,"native","adult",15,9,6)
}
