% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_whole.R
\name{sim_whole}
\alias{sim_whole}
\title{Simulate data and run model}
\usage{
sim_whole(
  sensitivity_analysis = FALSE,
  n_participants,
  change_coeff,
  change_coeff_value,
  check_convergence = FALSE,
  effect_of_interest,
  tau_0 = 0.9
)
}
\arguments{
\item{sensitivity_analysis}{Logical. Allows the number of participants and value of one coefficient to be manually overridden if TRUE. "sensitivity_analysis = FALSE" by default.}

\item{n_participants}{Numeric argument. Only needed if "sensitivity_analysis = TRUE". Specify total number of participants. Must be divisible by the number of groups (4 in this case) without a remainder.}

\item{change_coeff}{String argument. Only needed if "sensitivity_analysis = TRUE". Specifies the changed coefficient for the sensitivity analysis. E.g. "b0","b1". NOT "age" or "age_dev".}

\item{change_coeff_value}{Numeric argument. Only needed if "sensitivity_analysis = TRUE". Value of changed coefficient}

\item{check_convergence}{Logical. Check convergence if TRUE. "check_convergence = FALSE" by default.}

\item{effect_of_interest}{String argument. Only needed if "sensitivity_analysis = TRUE".}
}
\value{
"sensitivity_analysis = FALSE: \code{tibble} containing the model outputs of the fixed effects. "sensitivity_analysis = TRUE": return power (numeric argument)
}
\description{
Simulate data and run model
}
\examples{
# Normal version
sim_whole()

# Sensitivity analysis

sim_whole(sensitivity_analysis = TRUE,
n_participants = 60, # 60 participants in total (15 in each condition)
change_coef = "b1", # change "b2" (language coef)
change_coef_value = 90 # change "b2" to 90.
)

}
